io: refactor string and file work into new unit
